+++ js/main.js
@@ -1,26 +1,5 @@
-const descriptions = [
-  'Весело',
-  'Осень',
-  'На учебе',
-  'скучно',
-];
-const names = [
-  'Иван',
-  'Марк',
-  'Евгений',
-  'Мария',
-  'Ангелина',
-  'Илья',
-  'Вера',
-];
-const textComments = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
+import {descriptions, names, textComments} from './utils/data.js';
+import {getRandomInt, checkMaxStringLength} from './utils/util.js';
 
 const MIN_DESCRIPTION_IDENTIFIER = 1;
 const MAX_DESCRIPTION_IDENTIFIER = 25;
@@ -30,22 +9,6 @@ const MAX_NUMBER_OF_LIKES = 200;
 
 const MIN_NUMBER_FOR_LIKES = 1;
 const MAX_NUMBER_FOR_LIKES = 6;
-
-function getRandomInt(min, max) {
-  if (min >= 0 && max >= 0 && max >= min) {
-    const minUpdate = Math.ceil(min);
-    const maxUpdate = Math.floor(max);
-    return Math.floor(Math.random() * (maxUpdate - minUpdate)) + minUpdate;
-  } else {
-    return null;
-  }
-}
-
-function checkMaxStringLength(checkedString, maxLength) {
-  return checkedString.length <= maxLength;
-}
-
-checkMaxStringLength('Какой-то тестовый комментарий', 140);
 
 const getRandomComment = () => ({
   id: getRandomInt(MIN_DESCRIPTION_IDENTIFIER, MAX_DESCRIPTION_IDENTIFIER),
@@ -64,3 +27,4 @@ const getRandomPhoto = () => ({
 
 getRandomComment();
 getRandomPhoto();
+checkMaxStringLength('Какой-то тестовый комментарий', 140);
