+++ js/utils/data.js
@@ -1,0 +1,25 @@
+const descriptions = [
+  'Весело',
+  'Осень',
+  'На учебе',
+  'скучно',
+];
+const names = [
+  'Иван',
+  'Марк',
+  'Евгений',
+  'Мария',
+  'Ангелина',
+  'Илья',
+  'Вера',
+];
+const textComments = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {descriptions, names, textComments};
